rename schema file to reflect it's contents
